feat(player): add deinterlacing options for the Lume Engine
