+++ Lume/Services/Player/PlayerSettings.swift
@@ -296,6 +296,8 @@ enum PlayerSettings {
     /// everything else.
     enum Lume {
         static let hardwareDecodeKey = "player.lume.hardwareDecode"
+        static let deinterlaceModeKey = "player.lume.deinterlaceMode"
+        static let deinterlaceRateKey = "player.lume.deinterlaceRate"
         static let httpReconnectKey = "player.lume.httpReconnect"
         static let liveBufferKey = "player.lume.liveBuffer"
         static let vodBufferKey = "player.lume.vodBuffer"
@@ -324,7 +326,8 @@ enum PlayerSettings {
         /// values so each `@AppStorage` binding reverts to its default.
         static var allKeys: [String] {
             [
-                hardwareDecodeKey, httpReconnectKey, liveBufferKey, vodBufferKey,
+                hardwareDecodeKey, deinterlaceModeKey, deinterlaceRateKey,
+                httpReconnectKey, liveBufferKey, vodBufferKey,
                 videoQueueDepthKey, audioQueueDepthKey, stallThresholdKey,
                 ioTimeoutKey, probeSizeKey, analyzeDurationKey
             ]
